requirements.txt
python-telegram-bot==20.7
pandas
numpy
matplotlib
scipy
requests
flask
